Moved source to src/ and configured build task
