+++ .babelrc
@@ -1,0 +1,6 @@
+{
+    "presets": [
+        "es2015",
+        "stage-2"
+    ]
+}
