+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 /data
 /package-lock.json
 /config.json
+/lib
