+++ docker-compose.lowdb.yml
@@ -9,4 +9,4 @@ services:
     volumes:
       - ./:/app
     working_dir: /app
-    command: npm start
+    command: npm run start:dev
