+++ docker-compose.redis.yml
@@ -14,4 +14,4 @@ services:
     volumes:
       - ./:/app
     working_dir: /app
-    command: npm start
+    command: npm run start:dev
